dns_server/pyproject.toml
[project]
name = "dns-server"
version = "0.1.0"
description = "DNS server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "rsa>=4.9.1"
]

[dependency-groups]
dev = [
    "pytest-benchmark[histogram]>=5.2.3",
]